feat(hal): rcc clock configure
